Interaction can be chosen
The user now has the option to choose the interaction in all() such that
the data is grouped based on the chosen interaction. If the user chooses
"all", there is no grouping and the indicator is computed for all
interaction types. If the user chooses e.g. [["call","text"],"physical"]
or ["callandtext","physical] or other combinations, the indicator is
computed for both "call" and "text" and "physical" separately.